refactor: change settings directory
